Convert AlKharidGate plugin from kts to kt
